refactor(web): refactor sidebar header
